requirements.txt
openai
requests
beautifulsoup4
python-dotenv
ddgs
pypdf2
streamlit